requirements.txt
pandas
sqlalchemy
psycopg2-binary
apache-airflow